requirements.txt
Package           Version
----------------- -----------
amqp              5.2.0
annotated-types   0.7.0
anyio             4.4.0
billiard          4.2.0
celery            5.4.0
certifi           2024.6.2
click             8.1.7
click-didyoumean  0.3.1
click-plugins     1.1.1
click-repl        0.3.0
dnspython         2.6.1
email_validator   2.1.1
fastapi           0.111.0
fastapi-cli       0.0.4
h11               0.14.0
httpcore          1.0.5
httptools         0.6.1
httpx             0.27.0
idna              3.7
iniconfig         2.0.0
Jinja2            3.1.4
kombu             5.3.7
markdown-it-py    3.0.0
MarkupSafe        2.1.5
mdurl             0.1.2
motor             3.4.0
orjson            3.10.3
packaging         24.0
pip               23.2.1
pluggy            1.5.0
prompt_toolkit    3.0.46
psycopg           3.1.18
psycopg-pool      3.2.1
psycopg2          2.9.9
pydantic          2.7.3
pydantic_core     2.18.4
Pygments          2.18.0
pymongo           4.7.3
pytest            8.2.2
pytest-asyncio    0.23.7
python-dateutil   2.9.0.post0
python-dotenv     1.0.1
python-multipart  0.0.9
PyYAML            6.0.1
redis             5.0.5
rich              13.7.1
shellingham       1.5.4
six               1.16.0
sniffio           1.3.1
starlette         0.37.2
typer             0.12.3
typing_extensions 4.12.1
tzdata            2024.1
ujson             5.10.0
uvicorn           0.30.1
uvloop            0.19.0
vine              5.1.0
watchfiles        0.22.0
wcwidth           0.2.13
websockets        12.0
